requirements.txt
django~=5.0.3
requests
python-telegram-bot
